Move ffmpeg to its own mod
