+++ src/sources/mod.rs
@@ -1,1 +1,2 @@
+pub mod ffmpeg;
 pub mod youtube;
